Separate arrow from its editor
