+++ src/components/_shared/BezierArrow/BezierArrow.module.sass
@@ -1,0 +1,5 @@
+.arrow
+  stroke: currentColor
+  stroke-width: 2
+  stroke-linecap: round
+  fill: none
