+++ src/components/_shared/BezierArrow/BezierArrowEditor.js
@@ -1,10 +1,9 @@
 import { useRef, useState } from 'react'
-
 import { useDrag } from '@use-gesture/react'
 
-import _ from './BezierArrowEditor.module.sass'
+import BezierArrow from './BezierArrow'
 
-const LARGE_VALUE = 100000
+import _ from './BezierArrowEditor.module.sass'
 
 function DraggableCircle({ coords, handleDrag, className = '', ...rest }) {
   const ref = useRef()
@@ -20,35 +19,6 @@ function DraggableCircle({ coords, handleDrag, className = '', ...rest }) {
   )
 }
 
-function ArrowHead({ className, curvePath, coords, rotation, length }) {
-  return (
-    <g>
-      {[1, -1].map((direction) => (
-        <path
-          className={className}
-          transform={`rotate(${direction * rotation} ${coords.join(' ')})`}
-          style={{ strokeDasharray: [length, LARGE_VALUE].join(' ') }}
-          d={curvePath}
-        />
-      ))}
-    </g>
-  )
-}
-
-const constructCurve = (
-  startCoords,
-  endCoords,
-  startBezierHandle,
-  endBezierHandle
-) => [
-  'M',
-  `${startCoords}`,
-  'C',
-  `${startBezierHandle}`,
-  `${endBezierHandle}`,
-  `${endCoords}`
-]
-
 function BezierArrowEditor({
   initialStartCoords = [0, 0],
   initialEndCoords = [60, 60],
@@ -59,7 +29,8 @@ function BezierArrowEditor({
   drawArrowHead = true,
   arrowHeadAnchor = 'end', // one of 'start', 'end', 'both'
   arrowHeadLength = 10,
-  arrowHeadRotation = 30
+  arrowHeadRotation = 30,
+  className = '',
 }) {
   const [startCoords, setStartCoords] = useState(initialStartCoords)
   const [endCoords, setEndCoords] = useState(initialEndCoords)
@@ -73,52 +44,20 @@ function BezierArrowEditor({
     }
   }
 
-  const curve = constructCurve(startCoords, endCoords, startBezierHandle, endBezierHandle)
-  const invertCurve = constructCurve(endCoords, startCoords, endBezierHandle, startBezierHandle)
-
   return (
     <g className={_.editor}>
-      {/* Connect start/end points with bezier curve handles */}
-      <line
-        className={_.connector}
-        x1={startCoords[0]}
-        y1={startCoords[1]}
-        x2={startBezierHandle[0]}
-        y2={startBezierHandle[1]}
+      {/* Bezier Curve with arrow heads */}
+      <BezierArrow
+        className={className}
+        startCoords={startCoords}
+        endCoords={endCoords}
+        startBezierHandle={startBezierHandle}
+        endBezierHandle={endBezierHandle}
+        drawArrowHead={drawArrowHead}
+        arrowHeadAnchor={arrowHeadAnchor}
+        arrowHeadLength={arrowHeadLength}
+        arrowHeadRotation={arrowHeadRotation}
       />
-      <line
-        className={_.connector}
-        x1={endCoords[0]}
-        y1={endCoords[1]}
-        x2={endBezierHandle[0]}
-        y2={endBezierHandle[1]}
-      />
-
-      {/* Bezier curve */}
-      <path
-        className={_.curve}
-        d={curve.join(' ')}
-      />
-
-      {/* Arrow heads */}
-      {drawArrowHead && ['start', 'both'].includes(arrowHeadAnchor) && (
-        <ArrowHead
-          className={_.curve}
-          curvePath={curve.join(' ')}
-          coords={startCoords}
-          rotation={arrowHeadRotation}
-          length={arrowHeadLength}
-        />
-      )}
-      {drawArrowHead && ['end', 'both'].includes(arrowHeadAnchor) && (
-        <ArrowHead
-          className={_.curve}
-          curvePath={invertCurve.join(' ')}
-          coords={endCoords}
-          rotation={arrowHeadRotation}
-          length={arrowHeadLength}
-        />
-      )}
 
       {/* Start and end points */}
       <DraggableCircle
@@ -146,6 +85,22 @@ function BezierArrowEditor({
         coords={endBezierHandle}
         handleDrag={handleDrag(setEndBezierHandle)}
         r="6"
+      />
+
+      {/* Connect start/end points with bezier curve handles */}
+      <line
+        className={_.connector}
+        x1={startCoords[0]}
+        y1={startCoords[1]}
+        x2={startBezierHandle[0]}
+        y2={startBezierHandle[1]}
+      />
+      <line
+        className={_.connector}
+        x1={endCoords[0]}
+        y1={endCoords[1]}
+        x2={endBezierHandle[0]}
+        y2={endBezierHandle[1]}
       />
     </g>
   )
