+++ src/components/_shared/BezierArrow/BezierArrowEditor.module.sass
@@ -1,12 +1,6 @@
 .editor
   touch-action: 'none'
 
-  .curve
-    stroke: currentColor
-    stroke-width: 2
-    stroke-linecap: round
-    fill: none
-
   .connector
     stroke: currentColor
