+++ src/components/_shared/BezierArrow/utils.js
@@ -1,0 +1,13 @@
+export const constructCurve = (
+  startCoords,
+  endCoords,
+  startBezierHandle,
+  endBezierHandle
+) => [
+  'M',
+  `${startCoords}`,
+  'C',
+  `${startBezierHandle}`,
+  `${endBezierHandle}`,
+  `${endCoords}`
+]
